Refactored the codespace, splited the cli and restapi, added rest api file
