+++ main_test.go
@@ -1,0 +1,20 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"testing"
+)
+
+func TestEntryPoint(t *testing.T) {
+	fmt.Println(os.Args)
+
+	os.Args = []string{"oe"}
+	main()
+}
+
+func TestMain(m *testing.M) {
+	exitcode := m.Run()
+	os.RemoveAll("oe.log")
+	os.Exit(exitcode)
+}
